Move SharingIndexHandler into a separated file.
