+++ src/ContentRepository/SenseNet.ContentRepository.csproj
@@ -181,6 +181,7 @@
     <Compile Include="Packaging\Steps\SelectXmlValue.cs" />
     <Compile Include="Packaging\Steps\WhileDatabaseValue.cs" />
     <Compile Include="RepositoryBuilderExtensions.cs" />
+    <Compile Include="Search\Indexing\FieldIndexHandlers\SharingIndexHandler.cs" />
     <Compile Include="Search\Indexing\NullPopulator.cs" />
     <Compile Include="Sharing\SharingActions.cs" />
     <Compile Include="Sharing\SharingData.cs" />
